Altered MSR SVCR check to remove build warnings.

+++ src/lib/arch/aarch64/Instruction_execute.cc
@@ -3542,7 +3542,7 @@ void Instruction::execute() {
       case Opcode::AArch64_MSR: {  // msr (systemreg|Sop0_op1_Cn_Cm_op2), xt
         // TODO: Suppport reg writes to SVCR
         // Catch updates to SVCR - Not currently supported
-        if (metadata.operands[0].sys == ARM64_SYSREG_SVCR) {
+        if ((uint32_t)metadata.operands[0].sys == (uint32_t)ARM64_SYSREG_SVCR) {
           std::cerr << "[SimEng::Instruction_Execute] Register writes to SVCR "
                        "are not currently supported. Please use one of the "
                        "following :\n";
